+++ CHANGELOG.md
@@ -6,6 +6,7 @@ Most recent version is listed first.
 # v0.0.9
 - Use testdata in more testcases: https://github.com/komuw/kama/pull/46
 - Implement better identation: https://github.com/komuw/kama/pull/47
+- Pretty print context.Context: https://github.com/komuw/kama/pull/48
 
 # v0.0.8
 - Zero value struct should not indent: https://github.com/komuw/kama/pull/44